WIP Further attempts at defining wrapping functions
