--- src/shared/utils/call-wrappers.ts
@@ -1,77 +1,0 @@
-import { clearTimeout } from "timers";
-import { logger } from "./logger";
-
-const TIMEOUT_MESSAGE = "Operation timed out";
-export const ONLY_RETRY_ON_TIMEOUT = (error: Error): boolean =>
-  error.message === TIMEOUT_MESSAGE;
-
-export const callWithTimeout =
-  (timeoutMillis = 5000) =>
-  <TArgs, TResolution>(
-    asyncFunc: (underlyingArgs: TArgs) => Promise<TResolution>
-  ) =>
-  async (...underlyingArgs: any): Promise<TResolution> =>
-    await new Promise((resolve, reject) => {
-      const timeoutHandle = setTimeout(() => {
-        reject(new Error(TIMEOUT_MESSAGE));
-      }, timeoutMillis);
-      asyncFunc.apply(null, underlyingArgs).then(
-        (result) => {
-          clearTimeout(timeoutHandle);
-          resolve(result);
-        },
-        (error) => {
-          clearTimeout(timeoutHandle);
-          reject(error);
-        }
-      );
-    });
-
-export const callWithRetry =
-  (retries = 3, retryOnErrorMatching?: (error: Error) => boolean) =>
-  <TArgs, TResolution>(
-    asyncFunc: (underlyingArgs: TArgs) => Promise<TResolution>
-  ) =>
-  async (...underlyingArgs: any): Promise<TResolution> =>
-    await new Promise((resolve, reject) => {
-      let isUnresolved = true;
-      for (let i = 0; isUnresolved && i < retries; i++) {
-        asyncFunc.apply(null, underlyingArgs).then(
-          (result) => {
-            isUnresolved = false;
-            resolve(result);
-          },
-          (error) => {
-            if (
-              i < retries - 1 &&
-              (!retryOnErrorMatching || retryOnErrorMatching(error))
-            ) {
-              logger.warn(`Retrying on error: ${error.message}`);
-            } else {
-              reject(error);
-            }
-          }
-        );
-      }
-    });
-
-/* eslint-disable @typescript-eslint/ban-types */
-
-// This code is based on Redux Compose
-
-export const compose = (...funcs: Function[]): Function => {
-  if (funcs.length === 0) {
-    // infer the argument type, so it is usable in inference down the line
-    return <T>(arg: T) => arg;
-  }
-
-  if (funcs.length === 1) {
-    return funcs[0];
-  }
-
-  return funcs.reduce(
-    (a, b) =>
-      (...args: unknown[]) =>
-        a(b(...args))
-  );
-};
